BE/requirements.txt
Django>=5.0.4,<5.2
djangorestframework==3.14.0
psycopg[binary] >= 3.1.18
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
Pillow>=10.3.0
python-dotenv==1.0.0
drf-yasg>=1.21.7
django-filter==23.3
numpy
pandas
scikit-learn
